-Repair open-vm-tools port. -Update make.sh file.

+++ build/make.sh
@@ -3,15 +3,10 @@
 # Created: 2/12/2006 by Scott Zahn
 # Modified by Volker Theile (votdev@gmx.de)
 # Modified by Daisuke Aoyama (aoyama@peach.ne.jp)
+# Modified by Michael Zoon (michael.zoon@freenas.org)
 
 # Debug script
 #set -x
-
-# Exit if not managed error
-# http://www.davidpashley.com/articles/writing-robust-shell-scripts.html
-# "This tells bash that it should exit the script if any statement returns a non-true return value. 
-# The benefit of using -e is that it prevents errors snowballing into serious issues when they could have been caught earlier"
-set -e
 
 ################################################################################
 # Settings
@@ -25,7 +20,7 @@ FREENAS_SVNDIR="$FREENAS_ROOTDIR/svn"
 FREENAS_WORLD=""
 FREENAS_PRODUCTNAME=$(cat $FREENAS_SVNDIR/etc/prd.name)
 FREENAS_VERSION=$(cat $FREENAS_SVNDIR/etc/prd.version)
-FREENAS_REVISION=$(svnversion -n ${FREENAS_SVNDIR} | cut -d ":" -f1)
+FREENAS_REVISION=$(svn info ${FREENAS_SVNDIR} | grep "Revision:" | awk '{print $2}')
 FREENAS_ARCH=$(uname -p)
 FREENAS_KERNCONF="$(echo ${FREENAS_PRODUCTNAME} | tr '[:lower:]' '[:upper:]')-${FREENAS_ARCH}"
 FREENAS_OBJDIRPREFIX="/usr/obj/$(echo ${FREENAS_PRODUCTNAME} | tr '[:upper:]' '[:lower:]')"
@@ -197,27 +192,21 @@ build_world() {
 		fi
 
 		# Copy files from world.
-		for xfile in ${FREENAS_WORLD}/$file; do
-		    _file="${xfile#${FREENAS_WORLD}/}"
-		    if [ -e "$xfile" ]; then
-			if ! cp -fRpv $xfile $(echo $_file | rev | cut -d "/" -f 2- | rev) ; then
-			    echo "can't copy $xfile"
-			fi
-		    else
-			echo "WARNING, Missing file: $xfile"
-		    fi
-		done
+		cp -Rpv ${FREENAS_WORLD}/$file $(echo $file | rev | cut -d "/" -f 2- | rev)
 
 		# Deal with links
 		if [ $(echo "$i" | grep -c ":") -gt 0 ]; then
 			for j in $(echo $i | cut -d ":" -f 2- | sed "s/:/ /g"); do
-				ln -fsv /$file $j
+				ln -sv /$file $j
 			done
 		fi
 	done
 
 	# Cleanup
-	rm -f $FREENAS_WORKINGDIR/freenas.files
+	chflags -R noschg $FREENAS_TMPDIR
+	chflags -R noschg $FREENAS_ROOTFS
+	[ -d $FREENAS_TMPDIR ] && rm -f $FREENAS_WORKINGDIR/freenas.files
+	[ -f $FREENAS_WORKINGDIR/mfsroot.gz ] && rm -f $FREENAS_WORKINGDIR/mfsroot.gz
 
 	return 0
 }
@@ -234,9 +223,7 @@ create_rootfs() {
 	cp -v $FREENAS_SVNDIR/conf/config.xml .
 
 	# Compress zoneinfo data, exclude some useless files.
-	if [ ! -d $FREENAS_TMPDIR ] ; then
-		mkdir $FREENAS_TMPDIR
-	fi
+	mkdir $FREENAS_TMPDIR
 	echo "Factory" > $FREENAS_TMPDIR/zoneinfo.exlude
 	echo "posixrules" >> $FREENAS_TMPDIR/zoneinfo.exlude
 	echo "zone.tab" >> $FREENAS_TMPDIR/zoneinfo.exlude
@@ -354,9 +341,7 @@ add_libs() {
 	dirs=(${FREENAS_ROOTFS}/bin ${FREENAS_ROOTFS}/sbin ${FREENAS_ROOTFS}/usr/bin ${FREENAS_ROOTFS}/usr/sbin ${FREENAS_ROOTFS}/usr/local/bin ${FREENAS_ROOTFS}/usr/local/sbin ${FREENAS_ROOTFS}/usr/lib ${FREENAS_ROOTFS}/usr/local/lib ${FREENAS_ROOTFS}/usr/libexec ${FREENAS_ROOTFS}/usr/local/libexec)
 	for i in ${dirs[@]}; do
 		for file in $(find -L ${i} -type f -print); do
-            set +e
 			ldd -f "%p\n" ${file} 2> /dev/null >> /tmp/lib.list
-            set -e
 		done
 	done
 
@@ -372,7 +357,7 @@ add_libs() {
 	done
 
 	# Cleanup.
-	rm -f /tmp/lib.list || { echo "Can't rm /tmp/lib.list"; exit 1; } 
+	rm -f /tmp/lib.list
 
   return 0
 }
@@ -393,28 +378,23 @@ create_mfsroot() {
 	# Configure this file as a memory disk
 	md=`mdconfig -a -t vnode -f $FREENAS_WORKINGDIR/mfsroot`
 	# Create label on memory disk
-	#bsdlabel -m ${FREENAS_ARCH} -w ${md} auto || { echo "Can't bsdlabel the MFSroot md"; exit 1; } 
+	#bsdlabel -m ${FREENAS_ARCH} -w ${md} auto
 	# Format memory disk using UFS
-	newfs -S $FREENAS_IMGFMT_SECTOR -b $FREENAS_IMGFMT_BSIZE -f $FREENAS_IMGFMT_FSIZE -O2 -o space -m 0 /dev/${md} || { echo "Can't newfs the /dev/$(md)"; exit 1; } 
+	newfs -S $FREENAS_IMGFMT_SECTOR -b $FREENAS_IMGFMT_BSIZE -f $FREENAS_IMGFMT_FSIZE -O2 -o space -m 0 /dev/${md}
 	# Umount memory disk (if already used)
-    if [ `mount | grep -q $FREENAS_TMPDIR` ]; then
-        echo "Detected allready mounted workdir $FREENAS_TMPDIR"
-        umount $FREENAS_TMPDIR >/dev/null 2>&1
-    fi
+	umount $FREENAS_TMPDIR >/dev/null 2>&1
 	# Mount memory disk
-	mount /dev/${md} ${FREENAS_TMPDIR} || { echo "Can't mount /dev/$(md)";
- exit 1; }
+	mount /dev/${md} ${FREENAS_TMPDIR}
 	cd $FREENAS_TMPDIR
 	tar -cf - -C $FREENAS_ROOTFS ./ | tar -xvpf -
 
 	cd $FREENAS_WORKINGDIR
 	# Umount memory disk
-	umount $FREENAS_TMPDIR || { echo "Can't umount $FREENAS_TMPDIR"; exit 1; }
-	
-    # Detach memory disk
-	mdconfig -d -u ${md} || { echo "Cant' destroy ${md}"; exit 1; }
+	umount $FREENAS_TMPDIR
+	# Detach memory disk
+	mdconfig -d -u ${md}
 
-	gzip -9fnv $FREENAS_WORKINGDIR/mfsroot || { echo "Can't gzip mfsroot"; exit 1; }
+	gzip -9fnv $FREENAS_WORKINGDIR/mfsroot
 
 	return 0
 }
@@ -447,9 +427,7 @@ create_image() {
 	IMGFILENAME="${FREENAS_PRODUCTNAME}-${PLATFORM}-${FREENAS_VERSION}.${FREENAS_REVISION}.img"
 
 	echo "===> Generating tempory $FREENAS_TMPDIR folder"
-    if [ ! -d $FREENAS_TMPDIR ]; then
-	    mkdir $FREENAS_TMPDIR
-    fi
+	mkdir $FREENAS_TMPDIR
 	create_mfsroot;
 
 	echo "===> Creating an empty IMG file"
